pyproject.toml
[project]
name = "omnisupply-ai-multi-agent-supply-chain-intelligence-platform"
version = "0.1.0"
description = "Multi-agent AI system for supply chain intelligence with LangGraph and Opik observability"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.1.0",
    "langgraph>=1.0.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic[email]>=2.0.0",
    "python-dotenv>=1.2.1",
    "duckdb>=1.0.0",
    "plotly>=5.24.0",
    "scikit-learn>=1.5.0",
    "langchain-openai>=0.3.0",
    "opik>=0.2.0",
    "chromadb>=0.6.0",
    "tiktoken>=0.9.0",
    "nbformat>=5.10.4",
    "prophet>=1.2.1",
    "numpy>=2.3.5",
    "sqlalchemy>=2.0.0",
    "alembic>=1.17.2",
    "psycopg2-binary>=2.9.11",
    "fastapi>=0.123.5",
    "uvicorn[standard]>=0.38.0",
    "httpx>=0.28.1",
    "python-multipart>=0.0.20",
    "aiofiles>=25.1.0",
    "celery>=5.6.0",
    "redis>=7.1.0",
    "apscheduler>=3.11.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "black>=25.11.0",
    "ruff>=0.14.7",
    "duckdb-engine>=0.17.0",
    "email-validator>=2.3.0",
    "ipykernel>=6.29.0",
    "ipython>=8.12.0",
    "jupyter-client>=8.6.0",
    "openai>=1.0.0,<2.0.0",
]
